requirements.txt
discord.py>=2.4.0
aiohttp>=3.9.1
python-dotenv>=1.0.0
pandas>=2.1.4
numpy>=1.26.0
ta-lib
ccxt
websockets
scipy
asyncio-mqtt
pydantic
fastapi
uvicorn
redis
asyncpg
jinja2
python-telegram-bot
python-binance
colorlog
httpx
tenacity
rich
google-generativeai
schedule
google-generativeai==0.3.2
schedule==1.2.1